Translated lib/list/filter-util to typescript
